worker: Add GET parameter to skip the upstream fetch
When using Cloudflare worker sites it is not possible to assign the same
route to multiple workers. If we manually fire a request against
dashflare it will fail with an infinite loop because dashflare will try
to also fetch the original URL as part of the visit. This skips the
upstream fetching when a GET paremeter is present in the URL.