worker: Use the original URL when requests are forwarded
When the request is forwarded to the edge worker instead of running as a
transparent proxy (via the routes) the original URL gets lost. This PR
defines a new custom header `x-original-url` for passing the original
URL around. If the request is forwarded and the header is found it will
be used to extract the domain, path, query, etc. for the logging of the
visit.