worker: Use x-original-ip header if present
When forwarding a URL from a different worker (like when using a worker
site) Cloudflare will assign a different/local IP to the
`cf-connecting-ip` header which leads to wrong geolocation of the visit.
As we do for the handling of the original URL, if the `x-original-ip`
header is present its value will be used as the IP from the original
client.

Fixes #9