dashboard: Use the domain label in the queries
When the worker is deployed and requests handled via a `route` the
`host` and `domain` labels match. This is no the case where custom code
is needed for forwarding the request. This situation is particularly
important where dealing with Cloudflare's Worker Sites, since custom
code is needed for forward the URL and IP info the `host` will be the
subdomain where dashflare is deployed.