worker: Rename IPINFO_TOKEN to IPINFO
Since we switched geolocation providers a while back we no longer need a
token, we can either use the public available instance or selfhost a
private instance but in either of those cases a token is no longer
needed. Nevertheless, we keep the environment variable around as a way
of fully opting out of any geolocation of the users.